update(react): split useApi file
