+++ packages/react/src/api/useApi.ts
@@ -1,175 +1,10 @@
-import axios, { AxiosRequestConfig, AxiosResponse, CancelTokenSource } from 'axios'
-import { useCallback, useDebugValue, useEffect, useState } from 'react'
+import axios, { CancelTokenSource } from 'axios'
+import { useCallback, useDebugValue } from 'react'
 
-import { useCache } from './CacheContext'
-
-// Types
-export type Updator<R> = (data?: R) => R
-
-export type APIParams = Record<string, unknown>
-export type APIState<R> = { data?: R; loading: boolean }
-export type APIPromise<R> = Promise<R> & { cancel: () => void }
-
-export type APIGetRequestConfig = Omit<AxiosRequestConfig, 'cancelToken'> & { load?: boolean }
-export type APIGetRequestGenerator<R> = (source: CancelTokenSource) => Promise<AxiosResponse<R>>
-export type APIGetReturn<R> = APIState<R> & {
-  update: (data: R | Updator<R>) => void
-  reload: () => void
-}
-
-export type APIDeleteRequestGenerator<P extends APIParams, R> = (
-  source: CancelTokenSource,
-  url?: string,
-  params?: P
-) => Promise<AxiosResponse<R>>
-export type APIDeleteReturn<P extends APIParams, R> = APIState<R> & {
-  send: ((params?: P) => APIPromise<R>) & ((url: string, params?: P) => APIPromise<R>)
-}
-
-export type APIPostRequestConfig = Omit<AxiosRequestConfig, 'cancelToken'>
-export type APIPostRequestGenerator<B, P extends APIParams, R> = (
-  body: B,
-  source: CancelTokenSource,
-  params?: P
-) => Promise<AxiosResponse<R>>
-export type APIPostReturn<B, P extends APIParams, R> = APIState<R> & {
-  send: (data: B, params?: P) => APIPromise<R>
-}
-
-// Base hooks
-function useGetRequest<R>(generator: APIGetRequestGenerator<R>, cacheId: string, load = true): APIGetReturn<R> {
-  // Cache
-  const { data, setCache } = useCache<R>(cacheId)
-
-  // State
-  const [reload, setReload] = useState(load ? 1 : 0)
-  const [state, setState] = useState<APIState<R>>({ data, loading: true })
-
-  // Effect
-  useEffect(() => {
-    if (reload === 0) return
-    setState((old) => ({ ...old, loading: true }))
-
-    // Create cancel token
-    const source = axios.CancelToken.source()
-
-    // Make request
-    generator(source)
-      .then((res) => {
-        setState({ data: res.data, loading: false })
-      })
-      .catch((error) => {
-        if (axios.isCancel(error)) return
-
-        setState((old) => ({ ...old, loading: false }))
-        throw error
-      })
-
-    // Cancel
-    return () => {
-      source.cancel()
-    }
-  }, [generator, reload, setCache])
-
-  useEffect(() => {
-    if (state.data) setCache(state.data)
-  }, [state.data, setCache])
-
-  useEffect(() => {
-    setState((old) => ({ ...old, data }))
-  }, [cacheId])
-
-  return {
-    ...state,
-    update: useCallback(
-      (data: R | Updator<R>) => {
-        const updator: Updator<R> = typeof data === 'function' ? (data as Updator<R>) : () => data
-
-        setState((old) => ({ ...old, data: updator(old.data) }))
-      },
-      [setState]
-    ),
-    reload: useCallback(() => setReload((old) => old + 1), [setReload])
-  }
-}
-
-function useDeleteRequest<R, P extends APIParams>(generator: APIDeleteRequestGenerator<P, R>): APIDeleteReturn<P, R> {
-  // State
-  const [state, setState] = useState<APIState<R>>({ loading: false })
-
-  // Callback
-  const send = useCallback(
-    (arg1?: string | P, arg2?: P) => {
-      setState((old) => ({ ...old, loading: true }))
-
-      // Arguments
-      let url: string | undefined
-      let params: P | undefined
-
-      if (typeof arg1 === 'string') {
-        url = arg1
-        params = arg2
-      } else {
-        url = undefined
-        params = arg1
-      }
-
-      // Create cancel token
-      const source = axios.CancelToken.source()
-
-      // Make request
-      const promise = generator(source, url, params).then(
-        (res): R => {
-          setState({ data: res.data, loading: false })
-          return res.data
-        }
-      ) as APIPromise<R>
-
-      promise.cancel = () => source.cancel()
-      return promise
-    },
-    [generator]
-  )
-
-  return {
-    ...state,
-    send
-  }
-}
-
-function usePostRequest<B, R, P extends APIParams>(
-  generator: APIPostRequestGenerator<B, P, R>
-): APIPostReturn<B, P, R> {
-  // State
-  const [state, setState] = useState<APIState<R>>({ loading: false })
-
-  // Callback
-  const send = useCallback(
-    (body: B, params?: P) => {
-      setState((old) => ({ ...old, loading: true }))
-
-      // Create cancel token
-      const source = axios.CancelToken.source()
-
-      // Make request
-      const promise = generator(body, source, params).then(
-        (res): R => {
-          setState({ data: res.data, loading: false })
-          return res.data
-        }
-      ) as APIPromise<R>
-
-      promise.cancel = () => source.cancel()
-      return promise
-    },
-    [generator]
-  )
-
-  return {
-    ...state,
-    send
-  }
-}
+import { APIParams } from './types'
+import { APIDeleteReturn, useDeleteRequest } from './useDeleteRequest'
+import { APIGetRequestConfig, APIGetReturn, useGetRequest } from './useGetRequest'
+import { APIPostRequestConfig, APIPostReturn, usePostRequest } from './usePostRequest'
 
 // API Hooks
 export function useAPIGet<R, P extends APIParams = APIParams>(
@@ -232,8 +67,8 @@ export function useAPIDelete<R = unknown, P extends APIParams = APIParams>(
 
   // Callbacks
   const generator = useCallback(
-    (source: CancelTokenSource, _url?: string, _params?: P) =>
-      axios.delete<R>(_url || url, { ...config, params: { ...params, ..._params }, cancelToken: source.token }),
+    (source: CancelTokenSource, _params?: P) =>
+      axios.delete<R>(url, { ...config, params: { ...params, ..._params }, cancelToken: source.token }),
     [url, params, config]
   )
 
@@ -292,7 +127,7 @@ export function useAPIPatch<B, R = unknown, P extends APIParams = APIParams>(
 }
 
 // Namespaces
-const useAPI = {
+export const useAPI = {
   get: useAPIGet,
   delete: useAPIDelete,
   head: useAPIHead,
@@ -301,5 +136,3 @@ const useAPI = {
   put: useAPIPut,
   patch: useAPIPatch
 }
-
-export default useAPI
